requirements.txt
# python 3.6
nodeenv==1.2.0
# nodeenv -n 6.11.2 -p
django==1.9
djangorestframework==3.6.3
django-webpack-loader==0.5.0
channels
